requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-dotenv>=0.19.0
web3>=5.24.0
google-cloud-storage==2.14.0
python-magic==0.4.27
pillow==10.2.0
python-dateutil==2.8.2
cryptography==42.0.2
aiofiles>=0.7.0
jinja2>=3.0.1
python-socketio==5.11.1
itsdangerous==2.1.2
email-validator>=1.1.3 